+++ internal/bundler/bundler_default_test.go
@@ -4464,7 +4464,7 @@ func TestConstWithLet(t *testing.T) {
 			"/entry.js": `
 				const a = 1; console.log(a)
 				if (true) { const b = 2; console.log(b) }
-				if (true) { const b = 2; unknownFn(b) }
+				if (true) { const b = 3; unknownFn(b) }
 				for (const c = x;;) console.log(c)
 				for (const d in x) console.log(d)
 				for (const e of x) console.log(e)
@@ -4485,7 +4485,7 @@ func TestConstWithLetNoBundle(t *testing.T) {
 			"/entry.js": `
 				const a = 1; console.log(a)
 				if (true) { const b = 2; console.log(b) }
-				if (true) { const b = 2; unknownFn(b) }
+				if (true) { const b = 3; unknownFn(b) }
 				for (const c = x;;) console.log(c)
 				for (const d in x) console.log(d)
 				for (const e of x) console.log(e)